+++ src/producer/mod.rs
@@ -83,9 +83,11 @@ pub use self::base_producer::{
     DeliveryResult,
     DefaultProducerContext,
     ProducerContext,
-    ThreadedProducer
+    BaseRecord,
+    ThreadedProducer,
 };
 pub use self::future_producer::{
     FutureProducer,
-    DeliveryFuture
+    DeliveryFuture,
+    FutureRecord,
 };